Experiments in static scheduling (details follow)
It seems the openmp barrier is significantly slower than the work
being done. This is a particularly serious problem, since the
barrier is in shared memory. The best case here is that it's
in L3/LLC, but that involves a considerable amount of cache-coherency
overhead as barriers are constantly being waited on and released.
This not only slows down the program waiting for barrier-overhead,
but slows down actual X2/rowsum/Y reads, which are also coming from LLC.

Drastically increasing set/bin size is the only apparent solution here.
(other than going back to shuffling)